content-security-policy/to-port-from-blink/1.1/base-uri-allow.html

<!DOCTYPE html>
<html>
<head>
    <meta http-equiv="Content-Security-Policy" content="base-uri http://example.com">
    <base href="http://example.com/">
    <script src="http://localhost:8000/js-test-resources/js-test.js"></script>
    <script>
        description('Check that base URIs can be set if they do not violate the page\'s policy.');

        shouldBeEqualToString('document.baseURI', 'http://example.com/');
    </script>
</head>
<body>
</html>
